Jit: Followed unconditional branches are a property added by the Analyzer, not something to assume
If an unconditional branch is not the last instruction of the block, assuming that this means that the branch must have been followed is an unsafe assumption on how the analyzer works that may not hold true in the future.
As a bonus: JitArm64, unlike x86, didn't skip followed `bcx`s. Now it does.